requirements.txt
google-generativeai==0.5.2
langchain-google-genai==1.0.3
streamlit==1.36.0
python-dotenv==1.0.1
langchain==0.1.20
langchain-community==0.0.38
faiss-cpu==1.8.0
PyPDF2==3.0.1